Add separate logic to entity nodes
