+++ workspaces/ballerina/component-diagram/src/components/nodes/EntryNode/EntryNodeModel.ts
@@ -17,11 +17,17 @@
  */
 
 import { NodeModel } from "@projectstorm/react-diagrams";
+import { DiagramEngine } from "@projectstorm/react-diagrams-core";
 import { NodePortModel } from "../../NodePort";
 import { NODE_LOCKED, NodeTypes } from "../../../resources/constants";
 import { EntryPoint, EntryPointType } from "../../../utils/types";
 import { CDFunction, CDResourceFunction, CDService } from "@wso2/ballerina-core";
 import { getEntryNodeFunctionPortName } from "../../../utils/diagram";
+
+export interface BaseNodeWidgetProps {
+    model: EntryNodeModel;
+    engine: DiagramEngine;
+}
 
 export const VIEW_ALL_RESOURCES_PORT_NAME = "view-all-resources";
 export const GQL_GROUP_QUERY_PORT_NAME = "graphql-group-Query";
