+++ workspaces/ballerina/component-diagram/src/components/nodes/EntryNode/EntryNodeWidget.tsx
@@ -16,233 +16,13 @@
  * under the License.
  */
 
-import React, { useState } from "react";
-import styled from "@emotion/styled";
-import { DiagramEngine, PortWidget } from "@projectstorm/react-diagrams-core";
-import { EntryNodeModel, VIEW_ALL_RESOURCES_PORT_NAME, GQL_GROUP_QUERY_PORT_NAME, GQL_GROUP_MUTATION_PORT_NAME, GQL_GROUP_SUBSCRIPTION_PORT_NAME } from "./EntryNodeModel";
-import { NODE_BORDER_WIDTH, ENTRY_NODE_WIDTH, ENTRY_NODE_HEIGHT } from "../../../resources/constants";
-import { Button, Item, Menu, MenuItem, Popover, ImageWithFallback, ThemeColors, Icon, Codicon } from "@wso2/ui-toolkit";
-import { useDiagramContext } from "../../DiagramContext";
-import { HttpIcon, TaskIcon } from "../../../resources";
-import { MoreVertIcon } from "../../../resources/icons/nodes/MoreVertIcon";
-import { CDAutomation, CDFunction, CDService, CDResourceFunction } from "@wso2/ballerina-core";
-import { getEntryNodeFunctionPortName } from "../../../utils/diagram";
-import { GQLFuncListType, GQLState, PREVIEW_COUNT, SHOW_ALL_THRESHOLD } from "../../Diagram";
-type NodeStyleProp = {
-    hovered: boolean;
-    inactive?: boolean;
-};
-const Node = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    color: ${ThemeColors.ON_SURFACE};
-`;
-
-const Header = styled.div<NodeStyleProp>`
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
-    gap: 6px;
-    width: 100%;
-    cursor: pointer;
-`;
-
-const TopPortWidget = styled(PortWidget)`
-    margin-top: -3px;
-`;
-
-const BottomPortWidget = styled(PortWidget)`
-    margin-bottom: -2px;
-`;
-
-const FunctionPortWidget = styled(PortWidget)`
-`;
-
-const StyledText = styled.div`
-    font-size: 14px;
-`;
-
-const IconWrapper = styled.div`
-    padding: 4px;
-    max-width: 32px;
-    svg {
-        fill: ${ThemeColors.ON_SURFACE};
-    }
-    > div:first-child {
-        width: 24px;
-        height: 24px;
-        font-size: 24px;
-    }
-`;
-
-const Title = styled(StyledText) <NodeStyleProp>`
-    max-width: ${ENTRY_NODE_WIDTH - 80}px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    font-family: "GilmerMedium";
-    color: ${(props: NodeStyleProp) => (props.hovered ? ThemeColors.HIGHLIGHT : ThemeColors.ON_SURFACE)};
-    opacity: ${(props: NodeStyleProp) => (props.inactive && !props.hovered ? 0.7 : 1)};
-`;
-
-const ResourceAccessor = styled(StyledText) <{ color?: string }>`
-    text-transform: uppercase;
-    font-family: "GilmerBold";
-    background-color: ${(props) => props.color};
-    color: #FFF;
-    padding: 4px 8px;
-    border-radius: 4px;
-    font-size: 12px;
-    justify-content: center;
-    min-width: 60px;
-    text-align: center;
-    align-items: center;
-    font-weight: bold;
-`;
-
-const Description = styled(StyledText)`
-    font-size: 12px;
-    max-width: ${ENTRY_NODE_WIDTH - 80}px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    font-family: monospace;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-    color: ${ThemeColors.ON_SURFACE};
-    opacity: 0.7;
-`;
-
-const Box = styled.div<NodeStyleProp>`
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center;
-    gap: 8px;
-    width: 100%;
-
-    border: ${NODE_BORDER_WIDTH}px solid
-        ${(props: NodeStyleProp) => (props.hovered ? ThemeColors.HIGHLIGHT : ThemeColors.OUTLINE_VARIANT)};
-    border-radius: 8px;
-    background-color: ${ThemeColors.SURFACE_DIM};
-
-    padding: 0 8px 8px 8px;
-`;
-
-const ServiceBox = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    gap: 10px;
-    width: ${ENTRY_NODE_WIDTH}px;
-    height: ${ENTRY_NODE_HEIGHT}px;
-    cursor: pointer;
-`;
-
-const FunctionBoxWrapper = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    color: ${ThemeColors.ON_SURFACE};
-`;
-
-const StyledServiceBox = styled(ServiceBox) <NodeStyleProp>`
-    height: 40px;
-    padding: 0 12px;
-
-    border: ${NODE_BORDER_WIDTH}px solid
-        ${(props: NodeStyleProp) => (props.hovered ? ThemeColors.HIGHLIGHT : ThemeColors.OUTLINE_VARIANT)};
-    border-radius: 8px;
-    background-color: ${ThemeColors.SURFACE_DIM};
-`;
-
-// Visually group GraphQL operations with a colored accent bar
-const GroupContainer = styled.div<{ accent: string }>`
-    display: flex;
-    flex-direction: column;
-    gap: 6px;
-    width: ${ENTRY_NODE_WIDTH}px;
-    position: relative;
-    border: ${NODE_BORDER_WIDTH}px solid ${ThemeColors.OUTLINE_VARIANT};
-    border-radius: 8px;
-    background-color: ${ThemeColors.SURFACE_DIM};
-    padding: 6px;
-`;
-
-const GroupHeader = styled.div`
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    height: 40px;
-    padding: 0 6px 0 8px;
-    cursor: pointer;
-`;
-
-const MenuButton = styled(Button)`
-    border-radius: 5px;
-`;
-
-const ViewAllButton = styled(FunctionBoxWrapper)`
-    color: ${ThemeColors.PRIMARY};
-    height: 40px;
-    width: 100%;
-    cursor: pointer;
-    font-family: "GilmerMedium";
-    font-size: 14px;
-    &:hover {
-        border: 1px solid ${ThemeColors.HIGHLIGHT};
-        border-radius: 8px;
-    }
-`;
-
-const ViewAllButtonWrapper = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-`;
-
-const CollapseButton = styled(Button)`
-    border-radius: 5px;
-    padding: 2px;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 24px;
-    height: 24px;
-    min-width: 24px;
-    color: ${ThemeColors.ON_SURFACE};
-    & > div:first-of-type {
-        width: 16px;
-        height: 16px;
-        font-size: 16px;
-    }
-    svg {
-        fill: ${ThemeColors.ON_SURFACE} !important;
-    }
-    &:hover {
-        color: ${ThemeColors.HIGHLIGHT};
-    }
-    &:hover svg {
-        fill: ${ThemeColors.HIGHLIGHT} !important;
-    }
-`;
-
-const colors = {
-    "GET": '#3d7eff',
-    "PUT": '#fca130',
-    "POST": '#49cc90',
-    "DELETE": '#f93e3e',
-    "PATCH": '#986ee2',
-    "OPTIONS": '#0d5aa7',
-    "HEAD": '#9012fe'
-}
+import React from "react";
+import { DiagramEngine } from "@projectstorm/react-diagrams-core";
+import { EntryNodeModel } from "./EntryNodeModel";
+import { CDService } from "@wso2/ballerina-core";
+import { AIServiceWidget } from "./components/AIServiceWidget";
+import { GraphQLServiceWidget } from "./components/GraphQLServiceWidget";
+import { GeneralServiceWidget } from "./components/GeneralWidget";
 
 interface EntryNodeWidgetProps {
     model: EntryNodeModel;
@@ -253,463 +33,17 @@ export interface NodeWidgetProps extends Omit<EntryNodeWidgetProps, "children"> 
 
 export function EntryNodeWidget(props: EntryNodeWidgetProps) {
     const { model, engine } = props;
-    const [isHovered, setIsHovered] = React.useState(false);
-    const {
-        graphQLGroupOpen,
-        onServiceSelect,
-        onAutomationSelect,
-        onDeleteComponent,
-        onFunctionSelect,
-        expandedNodes,
-        onToggleNodeExpansion,
-        onToggleGraphQLGroup
-    } = useDiagramContext();
-    const [menuAnchorEl, setMenuAnchorEl] = useState<HTMLElement | SVGSVGElement>(null);
-    const isMenuOpen = Boolean(menuAnchorEl);
 
-
-    const handleOnClick = () => {
-        if (model.type === "service") {
-            if ((model.node as CDService).type === "ai:Service") {
-                onFunctionSelect(serviceFunctions[0]);
-            } else {
-                onServiceSelect(model.node as CDService);
-            }
-        } else {
-            onAutomationSelect(model.node as CDAutomation);
-        }
-    };
-
-    const handleToggleExpansion = (event: React.MouseEvent) => {
-        event.stopPropagation();
-        onToggleNodeExpansion(model.node.uuid);
-    };
-
-    const getNodeIcon = () => {
-        switch (model.type) {
-            case "automation":
-                return <TaskIcon />;
-            case "service":
-                const serviceType = (model.node as CDService)?.type;
-                const customIcon = getCustomEntryNodeIcon(serviceType);
-                if (customIcon) {
-                    return customIcon;
-                }
-                return <ImageWithFallback imageUrl={(model.node as CDService).icon} fallbackEl={<HttpIcon />} />;
-            default:
-                return <HttpIcon />;
-        }
-    };
-
-    const getNodeTitle = () => {
-        if (model.node.displayName) {
-            return model.node.displayName;
-        }
-        if ((model.node as CDService).absolutePath) {
-            return (model.node as CDService).absolutePath;
-        }
-        return "";
-    };
-
-    const getNodeDescription = () => {
-        if (model.type === "automation") {
-            return "Automation";
-        }
-        // Service
-        if ((model.node as CDService).type) {
-            return (model.node as CDService).type.replace(":Listener", ":Service");
-        }
-        return "Service";
-    };
-
-    const handleOnMenuClick = (event: React.MouseEvent<HTMLElement | SVGSVGElement>) => {
-        event.stopPropagation();
-        setMenuAnchorEl(event.currentTarget);
-    };
-
-    const handleOnMenuClose = () => {
-        setMenuAnchorEl(null);
-    };
-
-    const menuItems: Item[] = [
-        { id: "edit", label: "Edit", onClick: () => handleOnClick() },
-        { id: "delete", label: "Delete", onClick: () => onDeleteComponent(model.node) },
-    ];
-
-    const serviceFunctions = [];
-    if ((model.node as CDService).remoteFunctions?.length > 0) {
-        serviceFunctions.push(...(model.node as CDService).remoteFunctions);
-    }
-    if ((model.node as CDService).resourceFunctions?.length > 0) {
-        serviceFunctions.push(...(model.node as CDService).resourceFunctions);
-    }
-
-    const isExpanded = expandedNodes.has(model.node.uuid);
-    const hasMoreFunctions = serviceFunctions.length > 3;
-
-    let visibleFunctions;
-    if (serviceFunctions.length <= SHOW_ALL_THRESHOLD || isExpanded) {
-        visibleFunctions = serviceFunctions;
-    } else {
-        visibleFunctions = serviceFunctions.slice(0, PREVIEW_COUNT);
-    }
-
+    // Determine which widget to render based on service type
     if ((model.node as CDService)?.type === "ai:Service") {
-        return (
-            <Node>
-                <TopPortWidget port={model.getPort("in")!} engine={engine} />
-                <Box hovered={isHovered}>
-                    <ServiceBox
-                        onMouseEnter={() => setIsHovered(true)}
-                        onMouseLeave={() => setIsHovered(false)}
-                        onClick={() => { onFunctionSelect(serviceFunctions[0]) }}
-                    >
-                        <IconWrapper><Icon name="bi-ai-agent" /></IconWrapper>
-                        <Header hovered={isHovered} onClick={() => { onFunctionSelect(serviceFunctions[0]) }}>
-                            <Title hovered={isHovered}>{getNodeTitle().replace(/^\//, '')}</Title>
-                            <Description>{getNodeDescription()}</Description>
-                        </Header>
-                        <MenuButton appearance="icon" onClick={handleOnMenuClick}>
-                            <MoreVertIcon />
-                        </MenuButton>
-                    </ServiceBox>
-                </Box>
-                <Popover
-                    open={isMenuOpen}
-                    anchorEl={menuAnchorEl}
-                    handleClose={handleOnMenuClose}
-                    sx={{
-                        padding: 0,
-                        borderRadius: 0,
-                    }}
-                >
-                    <Menu>
-                        {menuItems.map((item) => (
-                            <MenuItem key={item.id} item={item} />
-                        ))}
-                    </Menu>
-                </Popover>
-                <BottomPortWidget port={model.getPort("out")!} engine={engine} />
-                <BottomPortWidget port={model.getPort(getEntryNodeFunctionPortName(serviceFunctions[0]))!} engine={engine} />
-            </Node>
-        );
+        return <AIServiceWidget model={model} engine={engine} />;
     }
 
-
-    return (
-        <Node>
-            <TopPortWidget port={model.getPort("in")!} engine={engine} />
-            <Box hovered={isHovered}>
-                <ServiceBox
-                    onMouseEnter={() => setIsHovered(true)}
-                    onMouseLeave={() => setIsHovered(false)}
-                    onClick={handleOnClick}
-                >
-                    <IconWrapper>{getNodeIcon()}</IconWrapper>
-                    <Header hovered={isHovered}>
-                        <Title hovered={isHovered}>{getNodeTitle()}</Title>
-                        <Description>{getNodeDescription()}</Description>
-                    </Header>
-                    <MenuButton appearance="icon" onClick={handleOnMenuClick}>
-                        <MoreVertIcon />
-                    </MenuButton>
-                </ServiceBox>
-                {(model.node as CDService)?.type === "graphql:Service" ? (
-                    <GraphQLGroups graphQLGroupOpen={graphQLGroupOpen} functions={serviceFunctions} model={model} engine={engine} onToggleGraphQLGroup={onToggleGraphQLGroup} />
-                ) : (
-                    <>
-                        {visibleFunctions.map((serviceFunction) => (
-                            <FunctionBox
-                                key={getEntryNodeFunctionPortName(serviceFunction)}
-                                func={serviceFunction}
-                                model={model}
-                                engine={engine}
-                            />
-                        ))}
-                        {hasMoreFunctions && !isExpanded && (
-                            <ViewAllButtonWrapper>
-                                <ViewAllButton onClick={handleToggleExpansion}>
-                                    Show More Resources
-                                </ViewAllButton>
-                                <PortWidget port={model.getPort(VIEW_ALL_RESOURCES_PORT_NAME)!} engine={engine} />
-                            </ViewAllButtonWrapper>
-                        )}
-                        {hasMoreFunctions && isExpanded && (
-                            <ViewAllButton onClick={handleToggleExpansion}>
-                                Show Fewer Resources
-                            </ViewAllButton>
-                        )}
-                    </>
-                )}
-            </Box>
-            <Popover
-                open={isMenuOpen}
-                anchorEl={menuAnchorEl}
-                handleClose={handleOnMenuClose}
-                sx={{
-                    padding: 0,
-                    borderRadius: 0,
-                }}
-            >
-                <Menu>
-                    {menuItems.map((item) => (
-                        <MenuItem key={item.id} item={item} />
-                    ))}
-                </Menu>
-            </Popover>
-            <BottomPortWidget port={model.getPort("out")!} engine={engine} />
-        </Node>
-    );
-}
-
-export function getColorByMethod(method: string) {
-    switch (method.toUpperCase()) {
-        case "GET":
-            return colors.GET;
-        case "PUT":
-            return colors.PUT;
-        case "POST":
-            return colors.POST;
-        case "DELETE":
-            return colors.DELETE;
-        case "PATCH":
-            return colors.PATCH;
-        case "OPTIONS":
-            return colors.OPTIONS;
-        case "HEAD":
-            return colors.HEAD;
-        default:
-            return '#876036'; // Default color
-    }
-}
-
-function FunctionBox(props: { func: CDFunction | CDResourceFunction; model: EntryNodeModel; engine: DiagramEngine, isGrouped?: boolean }) {
-    const { func, model, engine, isGrouped } = props;
-    const [isHovered, setIsHovered] = useState(false);
-    const { onFunctionSelect } = useDiagramContext();
-    const isGraphQL = (model.node as CDService)?.type === "graphql:Service";
-
-    const handleOnClick = () => {
-        onFunctionSelect(func);
-    };
-
-    const getAccessorDisplay = (accessor: string, isGraphQL: boolean): string => {
-        if (!isGraphQL) {
-            return accessor;
-        }
-
-        if (accessor === "get") return "Query";
-        if (accessor === "subscribe") return "Subscription";
-        return accessor;
-    };
-
-    return (
-        <FunctionBoxWrapper>
-            <StyledServiceBox
-                hovered={isHovered}
-                onClick={() => handleOnClick()}
-                onMouseEnter={() => setIsHovered(true)}
-                onMouseLeave={() => setIsHovered(false)}
-            >
-                {!isGrouped && (func as CDResourceFunction).accessor && (
-                    <ResourceAccessor color={getColorByMethod((func as CDResourceFunction).accessor)}>
-                        {getAccessorDisplay((func as CDResourceFunction).accessor, isGraphQL)}
-                    </ResourceAccessor>
-                )}
-                {(func as CDResourceFunction).path && (
-                    <Title hovered={isHovered}>
-                        {isGraphQL ? (func as CDResourceFunction).path : `/${(func as CDResourceFunction).path}`}
-                    </Title>
-                )}
-                {(func as CDFunction).name && <Title hovered={isHovered}>{(func as CDFunction).name}</Title>}
-            </StyledServiceBox>
-            <FunctionPortWidget port={model.getPort(getEntryNodeFunctionPortName(func))!} engine={engine} />
-        </FunctionBoxWrapper>
-    );
-}
-
-export function getCustomEntryNodeIcon(type: string) {
-    let typePart = type;
-    if (type && type.includes(":")) {
-        const typeParts = type.split(":");
-        typePart = typeParts.at(0);
+    if ((model.node as CDService)?.type === "graphql:Service") {
+        return <GraphQLServiceWidget model={model} engine={engine} />;
     }
 
-    switch (typePart) {
-        case "tcp":
-            return <Icon name="bi-tcp" />;
-        case "kafka":
-            return <Icon name="bi-kafka" />;
-        case "rabbitmq":
-            return <Icon name="bi-rabbitmq" sx={{ color: "#f60" }} />;
-        case "nats":
-            return <Icon name="bi-nats" />;
-        case "mqtt":
-            return <Icon name="bi-mqtt" sx={{ color: "#606" }} />;
-        case "grpc":
-            return <Icon name="bi-grpc" />;
-        case "graphql":
-            return <Icon name="bi-graphql" sx={{ color: "#e535ab" }} />;
-        case "java.jms":
-            return <Icon name="bi-java" />;
-        case "trigger.github":
-            return <Icon name="bi-github" />;
-        case "http":
-            return <Icon name="bi-globe" />;
-        default:
-            return null;
-    }
+    // Default to general service widget for all other cases
+    return <GeneralServiceWidget model={model} engine={engine} />;
 }
 
-export function GraphQLFunctionList(props: {
-    group: "Query" | "Subscription" | "Mutation";
-    functions: Array<CDFunction | CDResourceFunction>;
-    model: EntryNodeModel;
-    engine: DiagramEngine;
-    collapsed: boolean;
-    onToggleCollapse: (group: "Query" | "Subscription" | "Mutation") => void;
-}) {
-    const { group, functions, model, engine, collapsed, onToggleCollapse } = props;
-    const { expandedNodes, onToggleNodeExpansion } = useDiagramContext();
-
-    // using shared constants for preview and threshold
-
-    const accent = (() => {
-        switch (group) {
-            case "Query": return colors.GET;
-            case "Subscription": return colors.PUT;
-            case "Mutation": return colors.POST;
-        }
-    })();
-
-    const isExpanded = expandedNodes.has(model.node.uuid + group);
-    const canToggleItems = functions.length > SHOW_ALL_THRESHOLD;
-
-    let visibleFunctions;
-    if (functions.length <= SHOW_ALL_THRESHOLD || isExpanded) {
-        visibleFunctions = functions;
-    } else {
-        visibleFunctions = functions.slice(0, PREVIEW_COUNT);
-    }
-
-    const handleToggleExpansion = () => {
-        onToggleNodeExpansion(model.node.uuid + group);
-    };
-
-    return (
-        <FunctionBoxWrapper>
-            <GroupContainer accent={accent}>
-                <GroupHeader>
-                    <CollapseButton
-                        appearance="icon"
-                        aria-label={collapsed ? `Expand ${group}` : `Collapse ${group}`}
-                        onClick={(e: React.MouseEvent) => {
-                            e.stopPropagation();
-                            onToggleCollapse(group);
-                        }}
-                    >
-                        {collapsed ? (
-                            <Codicon name="chevron-right" />
-                        ) : (
-                            <Codicon name="chevron-down" />
-                        )}
-                    </CollapseButton>
-                    {group}
-                </GroupHeader>
-                {!collapsed && (
-                    <>
-                        {visibleFunctions.map((fn) => (
-                            <FunctionBox
-                                key={getEntryNodeFunctionPortName(fn)}
-                                func={fn}
-                                model={model}
-                                engine={engine}
-                                isGrouped
-                            />
-                        ))}
-                        {canToggleItems && !isExpanded && (
-                            <ViewAllButtonWrapper>
-                                <ViewAllButton onClick={handleToggleExpansion}>
-                                    Show More
-                                </ViewAllButton>
-                                <PortWidget
-                                    port={model.getPort(
-                                        group === "Query"
-                                            ? GQL_GROUP_QUERY_PORT_NAME
-                                            : group === "Mutation"
-                                                ? GQL_GROUP_MUTATION_PORT_NAME
-                                                : GQL_GROUP_SUBSCRIPTION_PORT_NAME
-                                    )!}
-                                    engine={engine}
-                                />
-                            </ViewAllButtonWrapper>
-                        )}
-                        {canToggleItems && isExpanded && (
-                            <ViewAllButton onClick={handleToggleExpansion}>
-                                Show Fewer
-                            </ViewAllButton>
-                        )}
-                    </>
-                )}
-            </GroupContainer>
-            {collapsed &&
-                <PortWidget
-                    port={model.getPort(
-                        group === "Query"
-                            ? GQL_GROUP_QUERY_PORT_NAME
-                            : group === "Mutation"
-                                ? GQL_GROUP_MUTATION_PORT_NAME
-                                : GQL_GROUP_SUBSCRIPTION_PORT_NAME
-                    )!}
-                    engine={engine}
-                >
-                </PortWidget>}
-        </FunctionBoxWrapper>
-    );
-}
-
-export function GraphQLGroups(props: { functions: Array<CDFunction | CDResourceFunction>; model: EntryNodeModel; engine: DiagramEngine; onToggleGraphQLGroup: (uuid: string, group: "Query" | "Subscription" | "Mutation") => void; graphQLGroupOpen: Record<string, GQLState> }) {
-    const { functions, model, engine, onToggleGraphQLGroup, graphQLGroupOpen } = props;
-
-    type GroupKey = "Query" | "Subscription" | "Mutation";
-
-    const getGroupLabel = (accessor?: string, name?: string): GroupKey | null => {
-        if (accessor === "get") return "Query";
-        if (accessor === "subscribe") return "Subscription";
-        if (!accessor && name) return "Mutation";
-        return null;
-    };
-
-    // Group functions by Query / Subscription / Mutation
-    const grouped: GQLFuncListType = functions.reduce((acc, fn) => {
-        const accessor = (fn as CDResourceFunction).accessor;
-        const name = (fn as CDFunction).name;
-        const group = getGroupLabel(accessor, name);
-        if (!group) return acc;
-        (acc[group] ||= []).push(fn);
-        return acc;
-    }, {} as GQLFuncListType);
-
-    const orderedGroups: GroupKey[] = ["Query", "Mutation", "Subscription"];
-
-    return (
-        <>
-            {orderedGroups
-                .filter((g) => grouped[g]?.length)
-                .map((group) => (
-                    <React.Fragment key={`gql-accordion-${group}`}>
-                        <GraphQLFunctionList
-                            group={group}
-                            functions={grouped[group]}
-                            model={model}
-                            engine={engine}
-                            collapsed={graphQLGroupOpen[model.node.uuid]?.[group] === false}
-                            onToggleCollapse={(group) => {
-                                onToggleGraphQLGroup(model.node.uuid, group);
-                            }}
-                        />
-                    </React.Fragment>
-                ))}
-        </>
-    );
-}
